+++ netbox_custom_objects/__init__.py
@@ -35,7 +35,7 @@ class CustomObjectsPluginConfig(PluginConfig):
     name = "netbox_custom_objects"
     verbose_name = "Custom Objects"
     description = "A plugin to manage custom objects in NetBox"
-    version = "0.4.5"
+    version = "0.4.6"
     author = 'Netbox Labs'
     author_email = 'support@netboxlabs.com'
     base_url = "custom-objects"
@@ -50,7 +50,7 @@ class CustomObjectsPluginConfig(PluginConfig):
     template_extensions = "template_content.template_extensions"
 
     @staticmethod
-    def _should_skip_dynamic_model_creation():
+    def should_skip_dynamic_model_creation():
         """
         Determine if dynamic model creation should be skipped.
 
@@ -143,7 +143,7 @@ class CustomObjectsPluginConfig(PluginConfig):
             )
 
             # Skip database calls if dynamic models can't be created yet
-            if self._should_skip_dynamic_model_creation():
+            if self.should_skip_dynamic_model_creation():
                 super().ready()
                 return
 
@@ -202,7 +202,7 @@ class CustomObjectsPluginConfig(PluginConfig):
             )
 
             # Skip custom object type model loading if dynamic models can't be created yet
-            if self._should_skip_dynamic_model_creation():
+            if self.should_skip_dynamic_model_creation():
                 return
 
             # Add custom object type models