+++ netbox_custom_objects/filtersets.py
@@ -83,6 +83,7 @@ class CustomObjectTypeFilterSet(NetBoxModelFilterSet):
         fields = (
             "id",
             "name",
+            "group_name",
         )