+++ netbox_custom_objects/models.py
@@ -200,6 +200,12 @@ class CustomObjectType(NetBoxModel):
     verbose_name = models.CharField(max_length=100, blank=True)
     verbose_name_plural = models.CharField(max_length=100, blank=True)
     slug = models.SlugField(max_length=100, unique=True, db_index=True, blank=False)
+    group_name = models.CharField(
+        max_length=100,
+        db_index=True,
+        blank=True,
+        help_text=_("Used to group similar custom object types in the navigation menu")
+    )
     cache_timestamp = models.DateTimeField(
         auto_now=True,
         help_text=_("Timestamp used for cache invalidation")
@@ -1714,6 +1720,10 @@ def clear_cache_on_field_save(sender, instance, **kwargs):
     """
     if instance.custom_object_type_id:
         CustomObjectType.clear_model_cache(instance.custom_object_type_id)
+    for pointing_field in CustomObjectTypeField.objects.filter(
+        related_object_type=instance.custom_object_type.object_type
+    ):
+        CustomObjectType.clear_model_cache(pointing_field.custom_object_type_id)
 
 
 @receiver(pre_delete, sender=CustomObjectTypeField)